LayoutTests/imported/w3c/web-platform-tests/css/css-values/lh-unit-005-expected.txt


FAIL lh relative to line-height inherited from parent assert_equals: expected "100px" but got "50px"
